requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dnspython==2.5.0
ecdsa==0.18.0
exceptiongroup==1.2.0
executing==2.0.1    
fastapi==0.109.0
filelock==3.9.0
flatbuffers==24.3.7
fonttools==4.47.2
fsspec==2023.4.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
ipykernel==6.29.0
ipython==8.20.0
jax==0.4.25
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.3.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.0
nodeenv==1.8.0
numpy==1.26.3
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
passlib==1.7.4
pillow==10.2.0
platformdirs==4.1.0
prisma==0.13.1
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.28.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
stack-data==0.6.3
starlette==0.35.1
StrEnum==0.4.15
sympy==1.12
threadpoolctl==3.2.0
tomlkit==0.12.4
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.13
uvicorn==0.27.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
